adding new views, urls, serializers

+++ core_api/views.py
@@ -23,10 +23,11 @@ class DepartmentsViews(APIView):
 
     def get(self, request, pk=None, format=None):
         if pk:
-            data = self.get_object(pk)
+            data = Departments.objects.get(pk = pk) #self.get_object(pk)
+            serializer = DepartmentsSerializer(data, many=False)
         else:
             data = Departments.objects.all()
-        serializer = DepartmentsSerializer(data, many=True)
+            serializer = DepartmentsSerializer(data, many=True)
         return Response(serializer.data)
 
     def post(self, request):
@@ -70,7 +71,7 @@ class DepartmentsViews(APIView):
 
 
 class SubDepartmentViews(APIView):
-    permission_classes = (IsAuthenticated,)
+    # permission_classes = (IsAuthenticated,)
     serializer_class = SubDepartmentSerializer
 
     def get_object(self, pk):
@@ -81,10 +82,11 @@ class SubDepartmentViews(APIView):
 
     def get(self, request, pk=None, format=None):
         if pk:
-            data = self.get_object(pk)
+            data = SubDepartments.objects.get(pk = pk) #self.get_object(pk)
+            serializer = SubDepartmentSerializer(data, many=False)
         else:
             data = SubDepartments.objects.all()
-        serializer = SubDepartmentSerializer(data, many=True)
+            serializer = SubDepartmentSerializer(data, many=True)
         return Response(serializer.data)
 
     def post(self, request):
@@ -130,6 +132,7 @@ class SubDepartmentViews(APIView):
 class DoctorViews(APIView):
     serializer_class = DoctorSerializer
 
+
     def get_object(self, pk):
         try:
             return Doctor.objects.filter(pk=pk)
@@ -138,10 +141,11 @@ class DoctorViews(APIView):
 
     def get(self, request, pk=None, format=None):
         if pk:
-            data = self.get_object(pk)
+            data = Doctor.objects.get(pk = pk) #self.get_object(pk)
+            serializer = DoctorSerializer(data, many=False)
         else:
             data = Doctor.objects.all()
-        serializer = DoctorSerializer(data, many=True)
+            serializer = DoctorSerializer(data, many=True)
         return Response(serializer.data)
 
     def post(self, request):
@@ -241,22 +245,6 @@ class ParameterViews(APIView):
             return Response(response, status=status.HTTP_400_BAD_REQUEST)
 
 
-class SampletypeViews(APIView):
-    serializer_class = SampletypeSerializer
-
-    def get_object(self, pk):
-        try:
-            return Sampletype.objects.filter(pk=pk)
-        except Sampletype.DoesNotExist:
-            raise Http404
-
-    def get(self, request, pk=None, format=None):
-        if pk:
-            data = self.get_object(pk)
-        else:
-            data = Sampletype.objects.all()
-        serializer = SampletypeSerializer(data, many=True)
-        return Response(serializer.data)
 
 class TitleViews(APIView):
     serializer_class = TitleSerializer
@@ -358,6 +346,63 @@ class ServiceViews(APIView):
                 return Response(response, status=status.HTTP_400_BAD_REQUEST)
             else:
                 instance = Service.objects.get(id=id)
+                instance.delete()
+                #### Below code for soft delete ##
+                # instance.is_active = False
+                # instance.save()
+                response = response_dict(
+                    data=dict(id=id), error=False, message="Successfully deleted.")
+                return Response(response, status=status.HTTP_200_OK)
+        except Exception as e:
+            response = response_dict(
+                data=None, error=True, message=str(e.__str__()))
+            return Response(response, status=status.HTTP_400_BAD_REQUEST)
+
+
+class SampleTypeViews(APIView):
+    serializer_class = SampletypeSerializer
+
+    def get_object(self, pk):
+        try:
+            return Sampletype.objects.filter(pk=pk)
+        except Sampletype.DoesNotExist:
+            raise Http404
+
+    def get(self, request, pk=None, format=None):
+        if pk:
+            data = self.get_object(pk)
+        else:
+            data = Sampletype.objects.all()
+        serializer = SampletypeSerializer(data, many=True)
+        return Response(serializer.data)
+
+    def post(self, request):
+        serializer = SampletypeSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response({"status": "success", "data": serializer.data}, status=status.HTTP_200_OK)
+        else:
+            return Response({"status": "error", "data": serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
+
+    def put(self, request, pk=None, format=None):
+        data = Sampletype.objects.get(pk=pk)
+        serializer = SampletypeSerializer(
+            instance=data, data=request.data, partial=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk, format=None):
+        try:
+            id = pk
+            if not (id and Sampletype.objects.filter(id=id).exists()):
+                response = response_dict(
+                    data=None, error=True, message="Not record found")
+                return Response(response, status=status.HTTP_400_BAD_REQUEST)
+            else:
+                instance = Sampletype.objects.get(id=id)
                 instance.delete()
                 #### Below code for soft delete ##
                 # instance.is_active = False
